requirements.txt
flask
flask-openapi3
flask-sqlalchemy
flask-migrate
flask-jwt-extended
pydantic[email]
python-dotenv